OAuth 2.0 code improvement
Improved codestyle, fix problems.
