+++ .env.example
@@ -1,0 +1,1 @@
+REACT_APP_API_URL = http://localhost:5000/api
