+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "dotenv": "^16.3.1",
     "nanoid": "^4.0.2",
     "node-sass": "8.0.0",
     "react": "^18.2.0",
