+++ src/redux/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
 
-import authReducer from "./auth-slice";
+import authReducer from "./slices/auth-slice";
 
 const store = configureStore({
     reducer: {
